Major overhaul of state and style
