+++ .gitignore
@@ -8,3 +8,5 @@ dist-ssr
 yarn.lock
 package-lock.json
 loot-assets
+character-assets
+.eslint-cache
