--- src/library/constants.js
@@ -1,6 +1,0 @@
-export default {
-  resetView: "resetView",
-  reactIsLoaded: "reactIsLoaded",
-  renderThumbnail: "renderThumbnail",
-  thumbnailResult: "thumbnailResult"
-};
