--- src/library/debug-config.js
@@ -1,6 +1,0 @@
-export default {
-  debugGLTF: true,
-  debugCanvases: false,
-  disableDownload: false,
-  debugExports: true,
-}
