--- src/styles/Selector.styled.js
@@ -1,3 +1,0 @@
-import styled from 'styled-components';
-import { SquareButton, ScreenSizeContainer } from './Globals.styled';
-
